Refactor widget management and layout components
- Removed unused imports and types from app.tsx
- Introduced new components: ColorConfigPanel, Draggable, Droppable, and PromptDialog for better modularity.
- Moved layout-related functions and types to dedicated files: layout-types.ts, layout-tree.ts, and layout-storage.ts.
- Updated widget catalog to maintain widget definitions and groups.
- Improved layout persistence with local storage handling.
