+++ src/app/lib/widget-catalog.tsx
@@ -1,0 +1,134 @@
+import {
+  ArrowRepeatAllRegular,
+  CalendarClockRegular,
+  ClockRegular,
+  FlagRegular,
+  GlobeRegular,
+  HistoryRegular,
+  HourglassRegular,
+  PlayCircleRegular,
+  PlugConnectedRegular,
+  PulseRegular,
+  RecordRegular,
+  TextFontRegular,
+} from "@fluentui/react-icons";
+import {
+  WidgetDefinition,
+  WidgetGroup,
+  WidgetKind,
+} from "./layout-types";
+
+export const WIDGET_GROUPS: { id: WidgetGroup; label: string }[] = [
+  { id: "clocks", label: "Clocks" },
+  { id: "playback", label: "Playback" },
+  { id: "timers", label: "Timers" },
+  { id: "ontime", label: "Ontime" },
+  { id: "recorders", label: "Recorders" },
+];
+
+export const widgetCatalog: Record<WidgetKind, WidgetDefinition> = {
+  worldClock: {
+    key: "worldClock",
+    label: "World Clock",
+    description: "Pick a configured timezone and pin it here",
+    icon: <GlobeRegular />,
+    color: "#3b82f6",
+    group: "clocks",
+  },
+  localClock: {
+    key: "localClock",
+    label: "Local Clock",
+    description: "Shows the local system time",
+    icon: <ClockRegular />,
+    color: "#14b8a6",
+    group: "clocks",
+  },
+  primaryTimer: {
+    key: "primaryTimer",
+    label: "Remaining Timer",
+    description: "Shows remaining time",
+    icon: <HourglassRegular />,
+    color: "#f59e0b",
+    group: "playback",
+  },
+  secondaryTimer: {
+    key: "secondaryTimer",
+    label: "Elapsed Timer",
+    description: "Shows elapsed time",
+    icon: <HistoryRegular />,
+    color: "#22c55e",
+    group: "playback",
+  },
+  loopState: {
+    key: "loopState",
+    label: "Loop State",
+    description: "Quick indicator for loop mode",
+    icon: <ArrowRepeatAllRegular />,
+    color: "#a855f7",
+    group: "playback",
+  },
+  timeOfDayCountdown: {
+    key: "timeOfDayCountdown",
+    label: "Time-of-day Countdown",
+    description: "Counts down to a target wall-clock time",
+    icon: <CalendarClockRegular />,
+    color: "#ef4444",
+    group: "timers",
+  },
+  oscTimer: {
+    key: "oscTimer",
+    label: "OSC Timer",
+    description: "Stopwatch triggered via OSC commands",
+    icon: <PulseRegular />,
+    color: "#ec4899",
+    group: "timers",
+  },
+  ontimeTimer: {
+    key: "ontimeTimer",
+    label: "Ontime Timer",
+    description: "Mirrors the current timer from Ontime over OSC",
+    icon: <PlugConnectedRegular />,
+    color: "#10b981",
+    group: "ontime",
+  },
+  ontimeTitle: {
+    key: "ontimeTitle",
+    label: "Ontime Title",
+    description: "Title of the currently running Ontime event",
+    icon: <TextFontRegular />,
+    color: "#0ea5e9",
+    group: "ontime",
+  },
+  ontimePlayback: {
+    key: "ontimePlayback",
+    label: "Ontime Playback",
+    description: "Ontime playback state (play, pause, stop, roll, armed)",
+    icon: <PlayCircleRegular />,
+    color: "#84cc16",
+    group: "ontime",
+  },
+  ontimeOnAir: {
+    key: "ontimeOnAir",
+    label: "Ontime On-Air",
+    description: "On-air indicator from Ontime",
+    icon: <RecordRegular />,
+    color: "#dc2626",
+    group: "ontime",
+  },
+  ontimeExpectedFinish: {
+    key: "ontimeExpectedFinish",
+    label: "Ontime Expected Finish",
+    description: "Wall-clock time the current Ontime event is expected to end",
+    icon: <FlagRegular />,
+    color: "#f97316",
+    group: "ontime",
+  },
+  recorderStatus: {
+    key: "recorderStatus",
+    label: "HyperDeck Recorder",
+    description: "Recording status of a Blackmagic HyperDeck",
+    icon: <RecordRegular />,
+    color: "#dc2626",
+    group: "recorders",
+  },
+};
